move Location and Language inside module/
