+++ src/main.rs
@@ -11,9 +11,7 @@ mod download;
 mod file_installer;
 mod get_module;
 mod global;
-mod language;
 mod list_components;
-mod location;
 mod log_parser;
 mod lowercase;
 mod module;
