+++ src/sub/invalidate.rs
@@ -1,7 +1,7 @@
 
 use crate::args::Invalidate;
 use crate::cache::Cache;
-use crate::location::{Location, Source};
+use crate::module::location::{Location, Source};
 use crate::lowercase::lwc;
 use crate::module::manifest::Manifest;
 use crate::module::module::Module;
